create View & ViewServices

+++ src/main/java/by/bulaukin/shop_receipt/countcost/countdiscount/CountingDiscount.java
@@ -4,5 +4,5 @@ import by.bulaukin.shop_receipt.countcost.CountingCost;
 
 public abstract class CountingDiscount extends CountingCost {
 
-    public abstract void setDiscount(int discount);
+    public abstract void setDiscount(Short discount);
 }
